+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 - the `tee` meta-filter can now forward or drop the incoming data based on a meta-data evaluation
 - added `sub-process` filter for processing data with sub-flow of filters, can be conditional based on meta-data evaluation
 - the `metadata-from-name` filter can work on the path now as well (must be present)
+- the `label-present-is` filter now lets the user to specify a minimum number of pixels that need to be present in the layers
 
 
 0.0.13 (2025-07-15)